{"schema":"decodex/commit/1","summary":"Modularize operator HTTP dashboard and APIs","authority":"manual"}
