+++ apps/decodex/src/orchestrator/operator_http.rs
@@ -3,13 +3,9 @@
 use std::{
 	io::{ErrorKind, Read, Write},
 	net::{TcpListener, TcpStream},
-	path::{Path, PathBuf},
-	sync::{
-		Arc, Mutex,
-		mpsc::{Receiver, RecvTimeoutError},
-	},
+	sync::{Arc, Mutex, mpsc::Receiver},
 	thread,
-	time::{Duration, Instant},
+	time::Duration,
 };
 
 use base64::engine::general_purpose::STANDARD;
@@ -30,13 +26,12 @@ use super::{
 	OPERATOR_ACCOUNTS_ENDPOINT_PATH, OPERATOR_APP_SNAPSHOT_ENDPOINT_PATH,
 	OPERATOR_DASHBOARD_ALIAS_ENDPOINT_PATH, OPERATOR_DASHBOARD_ENDPOINT_PATH,
 	OPERATOR_DASHBOARD_WS_CLIENT_MESSAGE_MAX_BYTES, OPERATOR_DASHBOARD_WS_ENDPOINT_PATH,
-	OPERATOR_DASHBOARD_WS_HEARTBEAT_INTERVAL, OPERATOR_LANE_INSPECT_ENDPOINT_PATH,
-	OPERATOR_LANE_INTERRUPT_ENDPOINT_PATH, OPERATOR_LANE_STEER_ALIAS_ENDPOINT_PATH,
-	OPERATOR_LANE_STEER_ENDPOINT_PATH, OPERATOR_LINEAR_SCAN_ENDPOINT_PATH,
-	OPERATOR_LIVE_ENDPOINT_PATH, OPERATOR_RUN_ACTIVITY_STREAM_INTERVAL,
-	OPERATOR_STATE_HEADER_TERMINATOR, OPERATOR_STATE_MAX_REQUEST_BYTES,
-	OperatorCodexAccountControlStatus, OperatorControlRequests, OperatorRunStatus,
-	OperatorStatusSnapshot, PublishedOperatorSnapshot,
+	OPERATOR_LANE_INSPECT_ENDPOINT_PATH, OPERATOR_LANE_INTERRUPT_ENDPOINT_PATH,
+	OPERATOR_LANE_STEER_ALIAS_ENDPOINT_PATH, OPERATOR_LANE_STEER_ENDPOINT_PATH,
+	OPERATOR_LINEAR_SCAN_ENDPOINT_PATH, OPERATOR_LIVE_ENDPOINT_PATH,
+	OPERATOR_RUN_ACTIVITY_STREAM_INTERVAL, OPERATOR_STATE_HEADER_TERMINATOR,
+	OPERATOR_STATE_MAX_REQUEST_BYTES, OperatorCodexAccountControlStatus, OperatorControlRequests,
+	OperatorRunStatus, OperatorStatusSnapshot, PublishedOperatorSnapshot,
 	build_operator_state_snapshot_without_live_observers, global_codex_account_control_status,
 	lane_control, operator_snapshot_presentation_value,
 };
