+++ apps/decodex/src/orchestrator/operator_http/dashboard/mod.rs
@@ -1,0 +1,20 @@
+mod connection;
+mod control;
+mod framing;
+mod handshake;
+mod run_activity;
+mod subscription;
+
+pub(crate) use self::run_activity::run_operator_run_activity_websocket_broadcasts;
+pub(super) use self::{
+	connection::handle_operator_dashboard_websocket_connection,
+	handshake::websocket_upgrade_required_response, subscription::dashboard_event_for_subscription,
+};
+
+#[cfg(test)]
+pub(crate) use self::{
+	framing::dashboard_websocket_message,
+	run_activity::{
+		build_operator_run_activity_event, strip_dashboard_run_activity_volatile_fields,
+	},
+};
